+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Feb 04 11:27:10 IST 2022
+gradle.version=7.3.3
