+++ src/test/java/com/mo/exercise/graph/DataUtilityTest.java
@@ -1,0 +1,40 @@
+package com.mo.exercise.graph;
+
+import junit.framework.TestCase;
+
+import java.util.List;
+
+public class DataUtilityTest extends TestCase
+{
+    public void testGetAllPeople()
+    {
+        DataUtility r = new DataUtility();
+        List<People> allPeople = r.getAllPeople();
+
+        assertEquals(12,allPeople.size());
+    }
+
+    public void testGetRelationShipCount()
+    {
+        DataUtility r = new DataUtility();
+        int bobRelationshipCount = r.getRelationShipCount("Bob");
+        int jennyRelationshipCount = r.getRelationShipCount("Jenny");
+        int nigelRelationshipCount = r.getRelationShipCount("Nigel");
+        int alanRelationshipCount = r.getRelationShipCount("Alan");
+
+        assertEquals(4,bobRelationshipCount);
+        assertEquals(3,jennyRelationshipCount);
+        assertEquals(2,nigelRelationshipCount);
+        assertEquals(0,alanRelationshipCount);
+    }
+
+    public void testGetExtendedFamilyCount()
+    {
+        DataUtility r = new DataUtility();
+        int bobExtendedFamilyCount = r.getExtendedFamilyCount("Bob");
+        int jennyExtendedFamilyCount = r.getExtendedFamilyCount("Jenny");
+
+        assertEquals(4,bobExtendedFamilyCount);
+        assertEquals(4,jennyExtendedFamilyCount);
+    }
+}
